Utility modules refactor + Added UE Sequencer related utility functions
