+++ Plugins/ueGear/Content/Python/ueGear/helpers.py
@@ -32,42 +32,6 @@ RELATIVE_PATH_PREFIX = './'
 BAD_RELATIVE_PATH_PREFIX = '../'
 WEB_PREFIX = 'https://'
 
-# Dictionary containing all the default material sample types
-MATERIAL_SAMPLER_TYPES = {
-	'Color': unreal.MaterialSamplerType.SAMPLERTYPE_COLOR,
-	'Grayscale': unreal.MaterialSamplerType.SAMPLERTYPE_GRAYSCALE,
-	'Alpha': unreal.MaterialSamplerType.SAMPLERTYPE_ALPHA,
-	'Normal': unreal.MaterialSamplerType.SAMPLERTYPE_NORMAL,
-	'Masks': unreal.MaterialSamplerType.SAMPLERTYPE_MASKS,
-	'DistanceFieldFont': unreal.MaterialSamplerType.SAMPLERTYPE_DISTANCE_FIELD_FONT,
-	'LinearColor': unreal.MaterialSamplerType.SAMPLERTYPE_LINEAR_COLOR
-}
-
-# Dictionary containing all the default Unreal Material properties
-MATERIAL_PROPERTIES = {
-	'BaseColor': unreal.MaterialProperty.MP_BASE_COLOR,
-	'Metallic': unreal.MaterialProperty.MP_METALLIC,
-	'Specular': unreal.MaterialProperty.MP_SPECULAR,
-	'Roughness': unreal.MaterialProperty.MP_ROUGHNESS,
-	'EmissiveColor': unreal.MaterialProperty.MP_EMISSIVE_COLOR,
-	'Normal': unreal.MaterialProperty.MP_NORMAL
-}
-
-# Dictionary containing default FBX import options
-DEFAULT_FBX_IMPORT_OPTIONS = {
-	'import_materials': True,
-	'import_textures': True,
-	'import_as_skeletal': False
-}
-
-# Dictionary containing default FBX export options
-DEFAULT_FBX_EXPORT_OPTIONS = {
-	'ascii': False,
-	'collision': False,
-	'level_of_detail': False,
-	'vertex_color': True
-}
-
 
 def is_python2():
 	"""
@@ -338,129 +302,6 @@ def get_game_world():
 	return unreal.get_editor_subsystem(unreal.UnrealEditorSubsystem).get_game_world()
 
 
-def get_all_actors_in_current_level():
-	"""
-	Returns all current actors in the current opened level.
-
-	:return: list of actors within level.
-	:rtype: list(unreal.Actor)
-	"""
-
-	return unreal.EditorActorSubsystem().get_all_level_actors()
-
-
-def get_selected_actors_in_current_level():
-	"""
-	Returns all current selected actors in the current opened level.
-
-	:return: list of selected actors within level.
-	:rtype: list(unreal.Actor)
-	"""
-
-	return unreal.EditorActorSubsystem().get_selected_level_actors()
-
-
-def get_all_actors_and_labels_in_current_level():
-	"""
-	Returns a tuple with all current actors and their labels in the current opened level.
-
-	:return: tuple with a list of actors and their labels within level.
-	:rtype: tuple(list(unreal.Actor), list(str))
-	"""
-
-	actors = get_all_actors_in_current_level()
-	actor_names = [actor.get_actor_label() for actor in actors]
-
-	return actors, actor_names
-
-
-def get_selected_actors_and_labels_in_current_level():
-	"""
-	Returns a tuple with all current selected actors and their labels in the current opened level.
-
-	:return: tuple with a list of selected actors and their labels within level.
-	:rtype: tuple(list(unreal.Actor), list(str))
-	"""
-
-	actors = get_selected_actors_in_current_level()
-	actor_names = [actor.get_actor_label() for actor in actors]
-
-	return actors, actor_names
-
-
-def get_all_actors_and_names_in_current_level():
-	"""
-	Returns a tuple with all current actors and their names in the current opened level.
-
-	:return: tuple with a list of actors and their names within level.
-	:rtype: tuple(list(unreal.Actor), list(str))
-	"""
-
-	actors = get_all_actors_in_current_level()
-	actor_names = [actor.get_name() for actor in actors]
-
-	return actors, actor_names
-
-
-def get_selected_actors_and_names_in_current_level():
-	"""
-	Returns a tuple with all current selected actors and their names in the current opened level.
-
-	:return: tuple with a list of selected actors and their names within level.
-	:rtype: tuple(list(unreal.Actor), list(str))
-	"""
-
-	actors = get_selected_actors_in_current_level()
-	actor_names = [actor.get_name() for actor in actors]
-
-	return actors, actor_names
-
-
-def get_actor_by_label_in_current_level(actor_label):
-	"""
-	Return actor within current level with the given actor label.
-
-	:param str actor_label: actor label.
-	:return: found actor with given label in current level.
-	:rtype: unreal.Actor or None
-
-	..warning:: Actor labels within a level are not unique, so if multiple actors have the same label we may return
-		the undesired one.
-	"""
-
-	found_actor = None
-	all_actors = get_all_actors_in_current_level()
-	for actor in all_actors:
-		if actor.get_actor_label() == actor_label:
-			found_actor = actor
-			break
-
-	return found_actor
-
-
-def get_all_actors_with_component_of_type(component_class):
-
-	found_actors = list()
-	actors = unreal.EditorActorSubsystem().get_all_level_actors()
-	for actor in actors:
-		components = actor.get_components_by_class(component_class)
-		if not components:
-			continue
-		found_actors.append(actor)
-
-	return found_actors
-
-
-def delete_actor(actor):
-	"""
-	Deletes given actor from current scene.
-
-	:param unreal.Actor actor: actor instance to delete.
-	"""
-
-	unreal.EditorLevelLibrary.destroy_actor(actor)
-
-
 def create_folder(root, name):
 	"""
 	Creates new folder.
@@ -480,585 +321,6 @@ def create_folder(root, name):
 		index += 1
 
 	return '{}/{}'.format(root, name)
-
-
-def list_asset_paths(directory='/Game', recursive=True, include_folder=False):
-	"""
-	Returns a list of all asset paths within Content Browser.
-
-	:param str directory: directory path of the asset we want the list from.
-	:param bool recursive: whether will be recursive and will look in sub folders.
-	:param bool include_folder: whether result will include folders name.
-	:param list(str) or str or None extra_paths: asset path of the asset.
-	:return: list of all asset paths found.
-	:rtype: list(str)
-	"""
-
-	return unreal.EditorAssetLibrary.list_assets(directory, recursive=recursive, include_folder=include_folder)
-
-
-def asset_exists(asset_path):
-	"""
-	Returns whether given asset path exists.
-
-	:param str asset_path: asset path of the asset.
-	:return: True if asset exist; False otherwise.
-	:rtype: bool
-	"""
-
-	return unreal.EditorAssetLibrary.does_asset_exist(asset_path)
-
-
-def get_asset_unique_name(asset_path, suffix=''):
-	"""
-	Returns a unique name for an asset in the given path.
-
-	:param str asset_path: path of the asset
-	:param str suffix: suffix to use to generate the unique asset name.
-	:return: tuple containing asset path and name.
-	:rtype: tuple(str, str)
-	"""
-
-	return unreal.AssetToolsHelpers.get_asset_tools().create_unique_asset_name(
-		base_package_name=asset_path, suffix=suffix)
-
-
-def rename_asset(asset_path, new_name):
-	"""
-	Renames asset with new given name.
-
-	:param str asset_path: path of the asset to rename.
-	:param str new_name: new asset name.
-	:return: new asset name.
-	:rtype: str
-	"""
-
-	dirname = os.path.dirname(asset_path)
-	new_name = dirname + '/' + new_name
-	unreal.EditorAssetLibrary.rename_asset(asset_path, new_name)
-	return new_name
-
-
-def move_assets_to_path(root, name, asset_paths):
-	"""
-	Moves/Rename the given list of assets to given destination directory.
-
-	:param str root: root of the path (eg. '/Game')
-	:param str name: name of the destination directory (eg. 'Target')
-	:param list(str) asset_paths: list of asset paths.
-	:return: new assets directory.
-	:rtype: str
-	"""
-
-	created_folder = create_folder(root, name)
-
-	for asset_path in asset_paths:
-		loaded = unreal.EditorAssetLibrary.load_asset(asset_path)
-		unreal.EditorAssetLibrary.rename_asset(asset_path, '{}/{}'.format(created_folder, loaded.get_name()))
-
-	return created_folder
-
-
-def get_assets(assets_path, recursive=False, only_on_disk=False):
-	"""
-	Returns all assets located in the given path.
-
-	:param str assets_path: path to get assets from.
-	:param bool recursive: whether to recursively find assets located in given path children folders.
-	:param bool only_on_disk: whether memory-objects will be ignored. If True, this function will be faster.
-	:return: assets data for all assets in the given path.
-	:rtype: list(unreal.AssetData)
-	"""
-
-	asset_registry = unreal.AssetRegistryHelpers.get_asset_registry()
-
-	return asset_registry.get_assets_by_path(
-		assets_path, recursive=recursive, include_only_on_disk_assets=only_on_disk) or list()
-
-
-def get_asset_data(asset_path, only_on_disk=False):
-	"""
-	Returns AssetData of the asset located in the given path.
-
-	:param str asset_path: path of the asset we want to retrieve data of.
-	:param bool only_on_disk: whether memory-objects will be ignored. If True, this function will be faster.
-	:return: data of the asset located in the given path.
-	:rtype: unreal.AssetData or None
-	"""
-
-	asset_registry = unreal.AssetRegistryHelpers.get_asset_registry()
-
-	return asset_registry.get_asset_by_object_path(asset_path, include_only_on_disk_assets=only_on_disk)
-
-
-def get_asset(asset_path, only_on_disk=False):
-	"""
-	Returns instance of an existent asset.
-
-	:param str asset_path: path of the asset instance we want to get.
-	:param bool only_on_disk: whether memory-objects will be ignored. If True, this function will be faster.
-	:return: instance of the asset located in the given path.
-	:rtype: object or None
-	"""
-
-	asset_data = get_asset_data(asset_path, only_on_disk=only_on_disk)
-	if not asset_data:
-		return None
-
-	full_name = asset_data.get_full_name()
-	path = full_name.split(' ')[-1]
-
-	return unreal.load_asset(path)
-
-
-def get_selected_asset_data():
-	"""
-	Returns current selected AssetData in Content Browser.
-
-	:return: list of selected asset data in Content Browser.
-	:rtype: list(AssetData)
-	"""
-
-	return unreal.EditorUtilityLibrary.get_selected_asset_data()
-
-
-def get_selected_assets():
-	"""
-	Returns current selected asset instances in Content Browser.
-
-	:return: list of selected asset instances in Content Browser.
-	:rtype: list(object)
-	"""
-
-	return unreal.EditorUtilityLibrary.get_selected_assets()
-
-
-def find_all_blueprints_data_assets_of_type(asset_type_name):
-	"""
-	Returns a list with all blueprint assets of the given type.
-
-	:param str or type asset_type_name: blueprint asset type name.
-	:return: list of blueprints assets with the given type.
-	:rtype: list
-	"""
-
-	found_blueprint_data_assets = list()
-	blueprints = unreal.AssetRegistryHelpers.get_asset_registry().get_assets_by_class('Blueprint', True)
-	for blueprint in blueprints:
-		blueprint_asset = blueprint.get_asset()
-		bp = unreal.EditorAssetLibrary.load_blueprint_class(blueprint_asset.get_path_name())
-		bp_type = unreal.get_type_from_class(bp)
-		if bp_type == asset_type_name or bp_type.__name__ == asset_type_name:
-			found_blueprint_data_assets.append(blueprint)
-
-	return found_blueprint_data_assets
-
-
-def create_asset(asset_path='', unique_name=True, asset_class=None, asset_factory=None, **kwargs):
-	"""
-	Creates a new Unreal asset.
-
-	:param str asset_path: path where the asset will be created.
-	:param bool unique_name: whether to automatically generate a unique name for the asset.
-	:param class asset_class: class of the asset we want to create.
-	:param class asset_factory: factory class to use for asset creation.
-	:param dict kwargs: custom keyword arguments to use by the asset creation factory.
-	:return: newly created asset instance.
-	:rtype: object or None
-	"""
-
-	if unique_name:
-		asset_path, asset_name = get_asset_unique_name(asset_path)
-	if not asset_exists(asset_path):
-		path = asset_path.rsplit('/', 1)[0]
-		name = asset_path.rsplit('/', 1)[1]
-		return unreal.AssetToolsHelpers.get_asset_tools().create_asset(
-			asset_name=name, package_path=path, asset_class=asset_class, factory=asset_factory, **kwargs)
-
-	return unreal.load_asset(asset_path)
-
-
-def import_assets(asset_tasks):
-	"""
-	Imports assets from the given asset import tasks.
-
-	:param list(unreal.AssetImportTask) asset_tasks: list of import tasks to run.
-	:return: list of imported asset paths.
-	:rtype: list(str)
-	"""
-
-	unreal.AssetToolsHelpers.get_asset_tools().import_asset_tasks(asset_tasks)
-
-	imported_paths = list()
-	for task in asset_tasks:
-		unreal.log('Import Task for: {}'.format(task.filename))
-		for object_path in task.imported_object_paths:
-			unreal.log('Imported object: {}'.format(object_path))
-			imported_paths.append(object_path)
-
-	return imported_paths
-
-
-def generate_fbx_import_task(
-		filename, destination_path, destination_name=None, replace_existing=True, automated=True, save=True,
-		fbx_options=None):
-	"""
-	Creates and configures an Unreal AssetImportTask to import a FBX file.
-
-	:param str filename: FBX file to import.
-	:param str destination_path: Content Browser path where the asset will be placed.
-	:param str or None destination_name: optional name of the imported asset. If not given, the name will be the
-		filename without the extension.
-	:param bool replace_existing: whether to replace existing assets.
-	:param bool automated: unattended import.
-	:param bool save: whether to save the file after importing it.
-	:param dict fbx_options: dictionary containing all the FBX settings to use.
-	:return: Unreal AssetImportTask that handles the import of the FBX file.
-	:rtype: unreal.AssetImportTask
-	"""
-
-	task = unreal.AssetImportTask()
-	task.filename = filename
-	task.destination_path = destination_path
-
-	# By default, task.destination_name is the filename without the extension
-	if destination_name:
-		task.destination_name = destination_name
-
-	task.replace_existing = replace_existing
-	task.automated = automated
-	task.save = save
-
-	task.options = unreal.FbxImportUI()
-	fbx_options = fbx_options or DEFAULT_FBX_IMPORT_OPTIONS
-
-	# Skeletal Mesh related import options
-	as_skeletal = fbx_options.pop('mesh_type_to_import', False)
-	skeletal_mesh_import_data = fbx_options.pop('skeletal_mesh_import_data', dict())
-	if skeletal_mesh_import_data:
-		sk_import_data = unreal.FbxSkeletalMeshImportData()
-		for name, value in skeletal_mesh_import_data.items():
-			try:
-				sk_import_data.set_editor_property(name, value)
-			except Exception:
-				unreal.log_warning('Was not possible to set Skeletal Mesh FBX Import property: {}: {}'.format(name, value))
-		task.options.skeletal_mesh_import_data = sk_import_data
-
-	# Base FBX import options
-	for name, value in fbx_options.items():
-		try:
-			task.options.set_editor_property(name, value)
-		except Exception:
-			unreal.log_warning('Was not possible to set FBX Import property: {}: {}'.format(name, value))
-	# task.options.static_mesh_import_data.combine_meshes = True
-
-	task.options.mesh_type_to_import = unreal.FBXImportType.FBXIT_STATIC_MESH
-	if as_skeletal:
-		task.options.mesh_type_to_import = unreal.FBXImportType.FBXIT_SKELETAL_MESH
-
-	return task
-
-
-def generate_fbx_export_task(asset, filename, replace_identical=True, automated=True, fbx_options=None):
-	"""
-	Creates and configures an Unreal AssetExportTask to export a FBX file.
-
-	:param str asset: asset we want to export with the task.
-	:param str filename: FBX file to export.
-	:param bool replace_identical: whether to replace identical files.
-	:param bool automated: unattended export.
-	:param fbx_options: dictionary containing all the FBX settings to use.
-	:return: Unreal AssetExportTask that handles the export of the FBX file.
-	:rtype: unreal.AssetExportTask
-	"""
-
-	task = unreal.AssetExportTask()
-	task.filename = filename
-	task.replace_identical = replace_identical
-	task.automated = automated
-	task.object = asset
-
-	task.options = unreal.FbxExportOption()
-	fbx_options = fbx_options or DEFAULT_FBX_EXPORT_OPTIONS
-	for name, value in fbx_options.items():
-		try:
-			task.options.set_editor_property(name, value)
-		except Exception:
-			unreal.log_warning('Was not possible to set FBX Export property: {}: {}'.format(name, value))
-
-	asset_class = asset.get_class()
-	exporter = None
-	if asset_class == unreal.StaticMesh.static_class():
-		exporter = unreal.StaticMeshExporterFBX()
-	elif asset_class == unreal.SkeletalMesh.static_class():
-		exporter = unreal.SkeletalMeshExporterFBX()
-	if not exporter:
-		unreal.log_warning('Asset Type "{}" has not a compatible exporter!'.format(asset_class))
-		return None
-	task.exporter = exporter
-
-	return task
-
-
-def import_fbx_asset(filename, destination_path, destination_name=None, import_options=None):
-	"""
-	Imports a FBX into Unreal Content Browser.
-
-	:param str filename: FBX file to import.
-	:param str destination_path: Content Browser path where the asset will be placed.
-	:param str or None destination_name: optional name of the imported asset. If not given, the name will be the
-		filename without the extension.
-	:param dict import_options: dictionary containing all the FBX import settings to use.
-	:return: path of the imported object.
-	:rtype: str
-	"""
-
-	tasks = list()
-	tasks.append(generate_fbx_import_task(
-		filename, destination_path, destination_name=destination_name, fbx_options=import_options))
-
-	return get_first_in_list(import_assets(tasks), default='')
-
-
-def export_fbx_asset(asset, directory, fbx_filename='', export_options=None):
-	"""
-	Exports a FBX from Unreal Content Browser.
-
-	:param unreal.Object asset: asset to export.
-	:param str directory: directory where FBX asset will be exported.
-	:param dict export_options: dictionary containing all the FBX export settings to use.
-	:return: exported FBX file path.
-	:rtype: str
-	"""
-
-	fbx_path = clean_path(os.path.join(directory, '{}.fbx'.format(fbx_filename or asset.get_name())))
-	export_task = generate_fbx_export_task(asset, fbx_path, fbx_options=export_options)
-	if not export_task:
-		unreal.log_warning('Was not possible to generate asset FBX export task')
-		return None
-
-	result = unreal.ExporterFBX.run_asset_export_task(export_task)
-
-	return fbx_path if result else ''
-
-
-def get_actor_asset(actor):
-	"""
-	Returns the asset associated to given actor based on actor components.
-
-	:param unreal.Actor actor: actor to get linked asset of.
-	:return: found asset.
-	:rtype: unreal.Object
-
-	..warning:: For now, only Static Mesh, Skeletal Meshes are supported.
-	"""
-
-	is_static_mesh = False
-	is_skeletal_mesh = False
-	is_camera = False
-	actor_asset = None
-
-	static_mesh_components = list()
-	skeletal_mesh_components = list()
-	camera_components = list()
-
-	static_mesh_components = actor.get_components_by_class(unreal.StaticMeshComponent)
-	if static_mesh_components:
-		is_static_mesh = True
-
-	if not is_static_mesh:
-		skeletal_mesh_components = actor.get_components_by_class(unreal.SkeletalMeshComponent)
-		if skeletal_mesh_components:
-			is_skeletal_mesh = True
-
-	if is_static_mesh:
-		static_mesh_component = get_first_in_list(static_mesh_components)
-		static_mesh = static_mesh_component.get_editor_property('static_mesh')
-		actor_asset = get_asset(static_mesh.get_path_name())
-	elif is_skeletal_mesh:
-		skeletal_mesh_component = get_first_in_list(skeletal_mesh_components)
-		skeletal_mesh = skeletal_mesh_component.get_editor_property('skeletal_mesh')
-		actor_asset = get_asset(skeletal_mesh.get_path_name())
-
-	return actor_asset
-
-
-def export_fbx_actor(actor, directory, export_options=None):
-	"""
-	Exports a FBX of the given actor.
-
-	:param unreal.Actor actor: actor to export as FBX.
-	:param str directory: directory where FBX actor will be exported.
-	:param dict export_options: dictionary containing all the FBX export settings to use.
-	:return: exported FBX file path.
-	:rtype: str
-	"""
-
-	actor_asset = get_actor_asset(actor)
-	if not actor_asset:
-		unreal.log_warning('Was not possible to retrieve valid asset for actor: {}'.format(actor))
-		return ''
-
-	return export_fbx_asset(
-		actor_asset, directory=directory, fbx_filename=actor.get_actor_label(), export_options=export_options)
-
-
-def export_all_fbx_actors_in_current_scene(directory, export_options=None):
-	"""
-	Exports all actors in current scene.
-
-	:param str directory: directory where FBX actor will be exported.
-	:param dict export_options: dictionary containing all the FBX export settings to use.
-	:return: exported FBX file paths.
-	:rtype: list(str)
-	"""
-
-	exported_actors = list()
-	all_actors = get_all_actors_in_current_level()
-	for actor in all_actors:
-		actor_export_path = export_fbx_actor(actor, directory=directory, export_options=export_options)
-		exported_actors.append(actor_export_path)
-
-	return exported_actors
-
-
-def generate_texture_import_task(
-		filename, destination_path, destination_name=None, replace_existing=True, automated=True, save=True,
-		import_options=None):
-	"""
-	Creates and configures an Unreal AssetImportTask to import a texture file.
-
-	:param str filename: texture file to import.
-	:param str destination_path: Content Browser path where the asset will be placed.
-	:param str or None destination_name: optional name of the imported texture. If not given, the name will be the
-		file name without the extension.
-	:param bool replace_existing:
-	:param bool automated:
-	:param bool save:
-	:param dict import_options: dictionary containing all the texture settings to use.
-	:return: Unreal AssetImportTask that handles the import of the texture file.
-	:rtype: unreal.AssetImportTask
-	"""
-
-	task = unreal.AssetImportTask()
-	task.filename = filename
-	task.destination_path = destination_path
-
-	# By default, task.destination_name is the filename without the extension
-	if destination_name:
-		task.destination_name = destination_name
-
-	task.replace_existing = replace_existing
-	task.automated = automated
-	task.save = save
-
-	texture_factory = unreal.TextureFactory()
-	task.factory = texture_factory
-	import_options = import_options or dict()
-	for name, value in import_options.items():
-		try:
-			task.factory.set_editor_property(name, value)
-		except Exception:
-			unreal.log_warning('Was not possible to set FBX Import property: {}: {}'.format(name, value))
-
-	return task
-
-
-def import_texture_asset(filename, destination_path, destination_name=None, import_options=None):
-	"""
-	Imports a texture asset into Unreal Content Browser.
-
-	:param str filename: texture file to import.
-	:param str destination_path: Content Browser path where the texture will be placed.
-	:param str or None destination_name: optional name of the imported texture. If not given, the name will be the
-		filename without the extension.
-	:param dict import_options: dictionary containing all the texture import settings to use.
-	:return: path to the imported texture asset.
-	:rtype: str
-	"""
-
-	tasks = list()
-	tasks.append(generate_texture_import_task(
-		filename, destination_path, destination_name=destination_name, import_options=import_options))
-
-	return get_first_in_list(import_assets(tasks), default='')
-
-
-def create_material(material_name, material_path, diffuse_color=None, roughness=None, specular=None):
-	"""
-	Creates a new Unreal material with given name.
-
-	:param str material_name: name of the material to create.
-	:param str material_path: path where material will be created project.
-	:param tuple(int, int) or None diffuse_color: default diffuse color for the material.
-	:param float or None roughness: default roughness for the material.
-	:param float or None specular: default specular for the material.
-	:return: newly created material.
-	:rtype: unreal.Material
-	"""
-
-	new_material = unreal.AssetToolsHelpers().get_asset_tools().create_asset(
-		material_name, material_path, unreal.Material, unreal.MaterialFactoryNew())
-
-	if diffuse_color is not None:
-		ts_node_diffuse = unreal.MaterialEditingLibrary.create_material_expression(
-			new_material, unreal.MaterialExpressionTextureSample, 0, 0)
-		for color_value, color_channel in zip(diffuse_color, 'RGB'):
-			ts_node_diffuse.set_editor_property(color_channel, color_value)
-		unreal.MaterialEditingLibrary.connect_material_property(
-			ts_node_diffuse, 'RGB', unreal.MaterialProperty.MP_BASE_COLOR)
-
-	if roughness is not None:
-		ts_node_roughness = unreal.MaterialEditingLibrary.create_material_expression(
-			new_material, unreal.MaterialExpressionConstant, 0, 0)
-		ts_node_roughness.set_editor_property('R', roughness)
-		unreal.MaterialEditingLibrary.connect_material_property(
-			ts_node_roughness, '', unreal.MaterialProperty.MP_ROUGHNESS)
-
-	if specular is not None:
-		ts_node_specular = unreal.MaterialEditingLibrary.create_material_expression(
-			new_material, unreal.MaterialExpressionConstant, 0, 0)
-		ts_node_specular.set_editor_property('R', roughness)
-		unreal.MaterialEditingLibrary.connect_material_property(
-			ts_node_specular, '', unreal.MaterialProperty.MP_SPECULAR)
-
-	return new_material
-
-
-def create_material_texture_sample(
-		material, pos_x=0, pos_y=0, texture=None, sampler_type=None, connect=True, from_expression='RGB',
-		property_to_connect=unreal.MaterialProperty.MP_BASE_COLOR):
-	"""
-	Creates a new material texture sample within the given material.
-
-	:param unreal.Material material: material we want to create texture sample for.
-	:param int pos_x:
-	:param int  pos_y:
-	:param texture:
-	:param sampler_type:
-	:param connect:
-	:param from_expression:
-	:param property_to_connect:
-	:return:
-	"""
-
-	material_texture_sample = unreal.MaterialEditingLibrary.create_material_expression(
-		material, unreal.MaterialExpressionTextureSample, pos_x, pos_y)
-	if texture:
-		material_texture_sample.set_editor_property('texture', texture)
-		if sampler_type is not None and sampler_type != '':
-			if is_string(sampler_type):
-				sampler_type = MATERIAL_SAMPLER_TYPES.get(sampler_type, None)
-			if sampler_type is not None:
-				material_texture_sample.set_editor_property('sampler_type', sampler_type)
-	if connect:
-		if is_string(property_to_connect):
-			property_to_connect = MATERIAL_PROPERTIES.get(property_to_connect, None)
-		if property_to_connect is not None and property_to_connect != '':
-			unreal.MaterialEditingLibrary.connect_material_property(
-				material_texture_sample, from_expression, property_to_connect)
-
-	return material_texture_sample
 
 
 def convert_maya_transforms_into_unreal_transforms(translation, rotation, scale):
@@ -1083,83 +345,12 @@ def convert_maya_transforms_into_unreal_transforms(translation, rotation, scale)
 	return unreal_translation, unreal_rotation, unreal_scale
 
 
-def remove_sequence_camera(level_sequence_name='', camera_name=''):
-	"""
-	Removes the camera from the given sequence with given name.
-
-	:param str level_sequence_name: name of the sequence that contains the camera.
-	:param str camera_name: name of the camera to remove.
-	:return: True if the camera was removed successfully; False otherwise.
-	:rtype: bool
-	"""
-
-	level_sequence_asset = unreal.load_asset(level_sequence_name)
-	if not level_sequence_asset:
-		return False
-
-	sequence_bindings = level_sequence_asset.get_bindings()
-	camera_bind = None
-	for binding in sequence_bindings:
-		if binding.get_name() or binding.get_display_name() == camera_name:
-			camera_bind = binding
-			break
-	if not camera_bind:
-		return False
-
-	unreal.LevelSequenceEditorBlueprintLibrary.close_level_sequence()
-	children_component = camera_bind.get_child_possessables() or list()
-	for child_component in children_component:
-		child_component.remove()
-	camera_tracks = camera_bind.get_tracks()
-	for camera_track in camera_tracks:
-		camera_bind.remove_track(camera_track)
-	camera_bind.remove()
-
-	unreal.LevelSequenceEditorBlueprintLibrary.refresh_current_level_sequence()
-	unreal.EditorAssetLibrary.save_loaded_asset(level_sequence_asset)
-
-	return True
-
-
-def get_subsequences(level_sequence_name):
-	"""
-	Returns a list of sequences from the given sequence name.
-
-	:param str level_sequence_name: name of the sequence whose subsequences we want to retrieve.
-	:return: list of sequence subsequences.
-	:rtype: list(unreal.LevelSequence)
-	"""
-
-	level_sequence_asset = unreal.load_asset(level_sequence_name)
-	if not level_sequence_asset:
-		return list()
-
-	found_subscene_track = None
-	tracks = level_sequence_asset.get_master_tracks()
-	for track in tracks:
-		if track.get_class() == unreal.MovieSceneSubTrack.static_class():
-			found_subscene_track = track
-			break
-
-	return found_subscene_track.get_sections() if found_subscene_track else list()
-
-
 def clear_level_selection():
 	"""
 	Clears the selection of the current opened level.
 	"""
 
 	unreal.EditorLevelLibrary().select_nothing()
-
-
-def select_actors_in_current_level(actors):
-	"""
-	Set the given actors as selected ones within current level.
-
-	:param unreal.Actor or list(unreal.Actor) actors: list of actors to select.
-	"""
-
-	unreal.EditorLevelLibrary().set_selected_level_actors(force_list(actors))
 
 
 def get_unreal_python_interpreter_path():
